backend_python/requirements.txt
Flask
python-dotenv
xarray
scikit-image
numpy
scipy
netCDF4
h5netcdf